Tech/requirements.txt
torch~=2.3.1
telebot~=0.0.5
openai~=1.68.2
Pillow~=9.5.0
torchvision~=0.18.1
pdfplumber~=0.11.5
pytesseract~=0.3.10